requirements.txt
flask
streamlit
numpy
pandas
langchain
llama-index
langchain-community
langchain-openai
langchain-core
fastapi
pydantic
gtts
uvicorn
transformers
sentencepiece
torch
sacremoses
-e .